Revise PropertyDecoratorFactory interface to support optional arguments.
 - Refactors composition logic out of types
 - Generalize property decorator factory
 - Map expected optional/mandatory options to decorator name
