+++ src/index.ts
@@ -1,6 +1,7 @@
 import 'reflect-metadata';
 
 export * from './api';
+export * from './composition';
 export * from './decorators';
 export * from './options';
 export * from './types';
