+++ src/types/property-decorator-factory.type.ts
@@ -1,3 +1,10 @@
-export interface PropertyDecoratorFactory<Options> {
+export type MandatoryPropertyDecoratorFactory<Options> = {
     (options: Options): PropertyDecorator;
-}
+};
+
+export type OptionalPropertyDecoratorFactory<Options> = {
+    (options?: Options): PropertyDecorator;
+};
+
+// eslint-disable-next-line max-len
+export type PropertyDecoratorFactory<Options> = MandatoryPropertyDecoratorFactory<Options> | OptionalPropertyDecoratorFactory<Options>;
